+++ src/main/java/com/example/demo/PrincipalController.java
@@ -12,12 +12,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
-@ComponentScan("com.system.model")
 public class PrincipalController {
-
-	@Autowired
-	private ApplicationContext context;
-
 	@GetMapping("/")
 	public String index() {
 		return "index";
@@ -27,22 +22,9 @@ public class PrincipalController {
 	public String Pagina1() {
 		return "pagina1";
 	}
+
 	@GetMapping("/p2")
 	public String Pagina2() {
 		return "pagina2";
 	}
-
-	@GetMapping("/p3")
-	public String contact(Model model) {
-		model.addAttribute("contato", new Contact(0,"","",""));
-		return "pagina3";
-	}
-
-	@PostMapping("/p3")
-	public String insertMessage(@ModelAttribute Contact c) {
-		ContactService cs = context.getBean(ContactService.class);
-		cs.insertContact(c);
-		return "pagina3";
-	}
-
 }
