+++ src/main/java/com/system/model/ContactService.java
@@ -1,7 +1,9 @@
 package com.system.model;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+@Service
 public class ContactService {
 
     @Autowired
