+++ src/main/resources/static/css/principal.css
@@ -2,12 +2,15 @@ body {
    background-color: #000;
 }
 
+h1{
+    text-align: center;
+}
 label{
-color: white;
+    color: white;
 }
 
 .center{
-text-align: center;
+    text-align: center;
 }
 
 nav.navbar-transparente {
