+++ src/main/resources/templates/pagina3.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
-    <meta charset="UTF-8">
+    <meta charset="utf-8">
 
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <link rel="stylesheet" th:href="@{css/principal.css}">
@@ -11,7 +11,7 @@
     <title>Fale Conosco</title>
 </head>
 
-<body  th:style="'background: url(/img/fd.png)'">
+<body th:style="'background: url(/img/fd.png)'">
 <header>
     <!--inicio cabecalho-->
     <nav class="navbar navbar-expand-md navbar-light fixed-top navbar-transparente">
@@ -47,11 +47,11 @@
 
 <section class="servicos">
     <div>
-        <form class="center" th:action="@{contact}" th:object="@{contact}" method="POST">
+        <form class="center" th:action="${contact}" method="POST" th:object="${contato}">
             <label for="name">Nome:</label><br>
             <input type="text" id="name" name="name" size="47" maxlength="40" required th:field="*{name}"><br>
             <label for="email">E-mail:</label><br>
-            <input type="email" id="email" name="email" size="47" maxlength="40" required th:field="*{email}"><br>
+            <input type="text" id="email" name="email" size="47" maxlength="40" required th:field="*{email}"><br>
             <label for="msg">Mensagem:</label><br>
             <textarea id="msg" name="msg" rows="5" cols="50" maxlength="500" th:field="*{msg}"></textarea><br><br>
