+++ src/main/resources/templates/success.html
@@ -11,7 +11,7 @@
     <title>Fale Conosco</title>
 </head>
 
-<body  th:style="'background: url(/img/fd.png)'">
+<body th:style="'background: url(/img/fd.png)'">
 <header>
     <!--inicio cabecalho-->
     <nav class="navbar navbar-expand-md navbar-light fixed-top navbar-transparente">
@@ -47,16 +47,7 @@
 
 <section class="servicos">
     <div>
-        <form class="center" th:action="${contact}" method="POST" th:object="${contato}">
-            <label for="name">Nome:</label><br>
-            <input type="text" id="name" name="name" size="47" maxlength="40" required th:field="*{name}"><br>
-            <label for="email">E-mail:</label><br>
-            <input type="text" id="email" name="email" size="47" maxlength="40" required th:field="*{email}"><br>
-            <label for="msg">Mensagem:</label><br>
-            <textarea id="msg" name="msg" rows="5" cols="50" maxlength="500" th:field="*{msg}"></textarea><br><br>
-
-            <input type="submit" value="Enviar">
-        </form>
+        <h1>Mensagem enviada com sucesso!</h1>
     </div>
 </section>
 </body>
